Moved files and made django changes
