+++ channel_models/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+class ChannelModelConfig(AppConfig):
+    name = 'channel_models'
+
+    def ready(self):
+        import channel_models.checklinks
