+++ channel_models/creators.py
@@ -1,11 +1,10 @@
-from youtubemissingvideonotify import checklinks
-from youtubemissingvideonotify import youtubegetter
 from youtubemissingvideonotify import keys
+from youtubemissingvideonotify.channel_models import checklinks, youtubegetter
 
 #### Setup
 channel_id = input("Whats your channel playlist ID?")
 youtubegetter.channels[channel_id] = {"email": "test@gmail.com"}
-videos = youtubegetter.fetch_videos(api_key=keys.key,channel_id=channel_id)
+videos = youtubegetter.fetch_videos(api_key=keys.key, channel_id=channel_id)
 youtubegetter.makeCSV(videos)
