+++ requirements.txt
@@ -1,0 +1,15 @@
+Django==1.11.4
+django-filter==1.0.4
+djangorestframework==3.6.3
+gunicorn==19.7.1
+Markdown==2.6.8
+pytz==2017.2
+redis==2.10.5
+Werkzeug==0.12.2
+requests
+py3
+psycopg2
+arrow
+huey
+django-extensions
+
